--- a/pyvista/core/dataset.py
+++ b/pyvista/core/dataset.py
@@ -202,6 +202,18 @@ class DataSet(DataSetFilters, DataObject):
     @points.setter
     def points(self, points: np.ndarray):
         """Set points without copying."""
+        if isinstance(points, pyvista_ndarray):
+            # simply set the underlying data
+            if points.VTKObject is not None:
+                try:
+                    self.GetPoints().SetData(points.VTKObject)
+                    self.GetPoints().Modified()
+                    self.Modified()
+                    return
+                except TypeError:
+                    pass
+
+        # otherwise, wrap and use the array
         if not isinstance(points, np.ndarray):
             raise TypeError('Points must be a numpy array')
         vtk_points = pyvista.vtk_points(points, False)