--- a/pyvista/utilities/geometric_objects.py
+++ b/pyvista/utilities/geometric_objects.py
@@ -50,7 +50,7 @@ def translate(surf, center=[0., 0., 0.], direction=[1., 0., 0.]):
 
     surf.transform(trans)
     if not np.allclose(center, [0., 0., 0.]):
-        surf.points[:] += np.array(center)
+        surf.points += np.array(center)
 
 
 def Cylinder(center=(0.0, 0.0, 0.0), direction=(1.0, 0.0, 0.0),