### Overview

When setting points inplace, we perform a shallow copy operation:
https://github.com/pyvista/pyvista/blob/6e61f8abff5f883bdc28234c39bb33c0b99ae12d/pyvista/core/dataset.py#L202-L214

Turns out, when the input array is a `pyvista_ndarray`, we keep around an extra reference, which was discovered in #1248 and #1247.  This PR resolves #1248 by simply setting the data directly from the pyvista_ndarray since it contains a reference to the original VTKObject.