@adeak 2021-04-14
The million-dollar question: where is the reference and why is it hanging around only some of the time (mostly with +=)?