@akaszynski 2021-04-14
> The million-dollar question: where is the reference and why is it hanging around only some of the time (mostly with +=)?

I think there's a circular reference.  Whenever an array is converted to a VTK object, we store the reference.  So the chain looks like:

With `mesh.points += <anything>` we use `@points.setter`.  This actually causes:
1. points = `mesh.points <pyvista_ndarray>`
2. mesh.points = points (calls `@points.setter`)
3. points is then converted to a `vtk` object, which includes a reference to the original `pyvista_ndarray`, which contains a reference to the original `VTKObject`, which was the original points object.  Hence why it's never collected.  Circular reference.

With mesh.points[:], we instead use `__setitem__` within the pyvista_ndarray itself.  No conversion using `numpy_to_vtk`.

---
Way forward is to remember that when we have a `pyvista_ndarray`, we have the original VTK object, and if we're writing to a vtk object, we should use VTK methods rather than going through an intermediate step.  This saves CPU cycles and eliminates these additional reference concerns.

This sounds complicated, and it is, but it's still better than the alternative: creating a copy of the data every time we convert to numpy.  It's speed + complexity vs simplicity.  Hard API choice.