@adeak 2021-04-14
OK, thanks, I think I'm starting to get it. So if I understand correctly the main issue was creating a new `vtkPoints` instance via `pv.vtk_points`.

So what if we take the last two lines of `vtk_points` when we can?
```py
    @points.setter
    def points(self, points: np.ndarray):
        """Set points without copying."""
        if not isinstance(points, np.ndarray):
            raise TypeError('Points must be a numpy array')
        pdata = self.GetPoints()
        if not pdata:
            # we're creating a new points array, create vtkPoints for it
            vtk_points = pyvista.vtk_points(points, False)
            self.SetPoints(vtk_points)
        else:
            # we're mutating existing points, only set the data
            pdata.SetData(_vtk.numpy_to_vtk(points, deep=True))
        self.GetPoints().Modified()
        self.Modified()
```

Or in diff format for readability:
```diff
    @points.setter
     def points(self, points: np.ndarray):
         """Set points without copying."""
         if not isinstance(points, np.ndarray):
             raise TypeError('Points must be a numpy array')
-        vtk_points = pyvista.vtk_points(points, False)
         pdata = self.GetPoints()
         if not pdata:
+            # we're creating a new points array, create vtkPoints for it
+            vtk_points = pyvista.vtk_points(points, False)
             self.SetPoints(vtk_points)
         else:
-            pdata.SetData(vtk_points.GetData())
+            # we're mutating existing points, only set the data
+            pdata.SetData(_vtk.numpy_to_vtk(points, deep=True))
         self.GetPoints().Modified()
         self.Modified()
```

Doing this change also seems to remove the gc issue at least in my local tests. And it might be a bit less complex.