@akaszynski 2021-04-14
> Doing this change also seems to remove the gc issue at least in my local tests. And it might be a bit less complex.

Your approach is indeed less complex.  We can implement that, but out of curiosity, is there any sort of performance improvement with mine vs yours?  I think `vtk_points` might do a copy somewhere.