@adeak 2021-04-14
>  is there any sort of performance improvement with mine vs yours? I think `vtk_points` might do a copy somewhere.

Good point, and I meant to add with my version that I didn't profile it in any way :) I have no idea.

But isn't `vtk_points()` only called once per mesh when the `.points` are being defined? That's how  I understood the `if not pdata:` case. Did I misunderstand that?