@adeak 2021-04-14
OK, your original version is so much faster that I think I'll just retract my suggestion.

Using this test:
```py
import numpy as np
import pyvista as pv

mesh = pv.Sphere().subdivide(4)
points_dummy = mesh.points.copy()

def set_points(N=1000):
    for _ in range(N):
        mesh.points = points_dummy

def update_points(N=1000):
    for _ in range(N):
        mesh.points += 0
```

Runtimes in ipython with my version:
```py
In [2]: %timeit set_points()
   ...: %timeit update_points()
188 ms ± 1.18 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
629 ms ± 22.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

And with the PR's version:
```py
In [2]: %timeit set_points()
   ...: %timeit update_points()
35.3 ms ± 387 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
133 ms ± 15 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```