@akaszynski 2021-04-15
Thanks for the benchmark!  This means that our setter is now way more performant.  What I don't like is how long `+=` takes (and shouldn't it take less time than set points?).  I'm going to look at this some more.