@adeak 2021-04-15
> This means that our setter is now way more performant.

Well I haven't benchmarked master, but it probably is :)

> and shouldn't it take less time than set points?

I don't think so. With `+=` you're creating a new `pyvista_ndarray` from scratch before using that to mutate the underlying data, whereas for the raw setting case the array to be used is already done. I wouldn't be surprised if the time difference was mostly the time it takes to compute `points + 0`.