@adeak 2021-04-15
Here's a bit more fair comparison where we're doing the same amount of work in the raw setter case:
```py
import numpy as np
import pyvista as pv

mesh = pv.Sphere().subdivide(4)  # 50k points
points_dummy = mesh.points.copy()

def set_points(N=1000):
    for _ in range(N):
        mesh.points = points_dummy + 0

def update_points(N=1000):
    for _ in range(N):
        mesh.points += 0

```

"Master", set vs. augmented assignment:
```
335 ms ± 8.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
177 ms ± 912 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This PR, set vs. augmented assignment:
```
496 ms ± 8.94 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
114 ms ± 681 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

My bad suggestion, set vs. augmented assignment:
```
472 ms ± 110 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
702 ms ± 117 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```