@akaszynski 2021-04-19
Seeing somewhat different timings:
```py
import numpy as np
import pyvista as pv

mesh = pv.Sphere().subdivide(4)  # 50k points
points_dummy = mesh.points.copy()

def set_points(N=1000):
    for _ in range(N):
        mesh.points = points_dummy + 0

def update_points(N=1000):
    for _ in range(N):
        mesh.points += 0


def update_points_inplace(N=1000):
    for _ in range(N):
        mesh.points[:] += 0
```

```

# master
 573 ms ± 30.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
 153 ms ± 4.48 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
 114 ms ± 903 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# this branch
 560 ms ± 3.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
 106 ms ± 762 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
 112 ms ± 374 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

```

Either way, this PR will improve garbage collection without the loss of performance.  Calling it good for today.